Feat/BE/Change : model into member and community

+++ BE/api/boards/views.py
@@ -3,6 +3,7 @@ from rest_framework.response import Response
 from rest_framework.decorators import api_view
 from rest_framework import status
 from .models import Board, BoardComment
+from community.models import Member
 from .serializers import BoardListSerializer, BoardSerializer, BoardCommentSerializer
 
 
@@ -17,12 +18,15 @@ def board_list(request):
         serializer = BoardSerializer(data=request.data)
 
         if serializer.is_valid(raise_exception=True):
-            serializer.save()
+            member = get_object_or_404(Member, user=request.user)
+            community = member.community
+            serializer.save(member=member, community=community)
             return Response(serializer.data, status=status.HTTP_201_CREATED)
 
 
 @api_view(['GET', 'DELETE', 'PUT'])
 def board_detail(request, board_pk):
+    member = get_object_or_404(Member, user=request.user)
     board = get_object_or_404(Board, pk=board_pk)
 
     if request.method == 'GET':
@@ -30,36 +34,42 @@ def board_detail(request, board_pk):
         return Response(serializer.data)
 
     elif request.method == 'DELETE':
-        board.delete()
-        return Response(status=status.HTTP_204_NO_CONTENT)
+        if member.user == request.user:
+            board.delete()
+            return Response(status=status.HTTP_204_NO_CONTENT)
 
     elif request.method == 'PUT':
-        serializer = BoardSerializer(board, data=request.data)
+        if member.user == request.user:
+            serializer = BoardSerializer(board, data=request.data)
 
-        if serializer.is_valid(raise_exception=True):
-            serializer.save()
-            return Response(serializer.data)
+            if serializer.is_valid(raise_exception=True):
+                serializer.save()
+                return Response(serializer.data)
 
 
 @api_view(['POST', 'DELETE', 'PUT'])
 def board_comment(request, board_pk, comment_pk=None):
+    member = get_object_or_404(Member, user=request.user)
+
     if request.method == 'POST':
         board = get_object_or_404(Board, pk=board_pk)
         serializer = BoardCommentSerializer(data=request.data)
 
         if serializer.is_valid(raise_exception=True):
-            serializer.save(board=board)
+            serializer.save(member=member, board=board)
             return Response(serializer.data, status=status.HTTP_201_CREATED)
 
     elif request.method == 'DELETE':
-        comment = get_object_or_404(BoardComment, pk=comment_pk)
-        comment.delete()
-        return Response(status=status.HTTP_204_NO_CONTENT)
+        if member.user == request.user:
+            comment = get_object_or_404(BoardComment, pk=comment_pk)
+            comment.delete()
+            return Response(status=status.HTTP_204_NO_CONTENT)
 
     elif request.method == 'PUT':
-        comment = get_object_or_404(BoardComment, pk=comment_pk)
-        serializer = BoardCommentSerializer(comment, data=request.data)
+        if member.user == request.user:
+            comment = get_object_or_404(BoardComment, pk=comment_pk)
+            serializer = BoardCommentSerializer(comment, data=request.data)
 
-        if serializer.is_valid(raise_exception=True):
-            serializer.save()
-            return Response(serializer.data)
+            if serializer.is_valid(raise_exception=True):
+                serializer.save()
+                return Response(serializer.data)
